requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.47.0
requests==2.32.3
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.0
dateparser==1.2.0
loguru==0.7.2
